Fix error when there are changes in objects of type "set"
Dexterity types that use selectable elements (zope.schema.Set) have
objects of type "set". When using CMFDiff tool to examine changes
between versions having differences between these set objects,
FieldDiff.inline_diff() fails as it assumes the objects a and b can be
indexed. This is solved by passing a list of the object, if the object
is of type "set".